pyproject.toml
[tool.poetry]
name = "motion"
version = "0.1.0"
description = ""
authors = ["Jeff Hutchison <jeff@jeffhutchison.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
robotpy-wpimath = "^2021.2.2"
jupyter = "^1.0.0"
matplotlib = "^3.4.1"
ipympl = "^0.7.0"
simple-pid = "^1.0.0"
more-itertools = "^8.7.0"
pandas = "^1.2.4"
robotpy = "^2021.2.0"
requests = "^2.25.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
